+++ lib/essentials/sidebar/application/sidebar_action_dispatcher.g.dart
@@ -7,7 +7,7 @@ part of 'sidebar_action_dispatcher.dart';
 // **************************************************************************
 
 String _$sidebarActionDispatcherHash() =>
-    r'cf05af522f64646799be6b70064a51dd7fda98b8';
+    r'3311e7dcf56fdfd0fbc998a9b292d397e58dfe60';
 
 /// See also [SidebarActionDispatcher].
 @ProviderFor(SidebarActionDispatcher)