+++ lib/features/settings/application/sidebar_cassette_spec/coordinators/settings_coordinator.dart
@@ -2,7 +2,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import '../../../../../essentials/sidebar/presentation/view_model/sidebar_cassette_card_view_model.dart';
 import '../../../../attachments/feature_level_providers.dart'
-    show AttachmentArchiveLocationState, AttachmentArchiveRelocationProgress;
+    show AttachmentArchiveLocationState;
 import '../../../domain/spec_classes/settings_cassette_spec.dart';
 import '../payloads/historical_archives_settings_cassette_payload.dart';
 import '../resolvers/attachment_archive_settings_resolver.dart';
@@ -24,8 +24,6 @@ class SettingsCassetteCoordinator extends _$SettingsCassetteCoordinator {
     required int cassetteIndex,
     List<HistoricalArchiveSidebarSourceSummary>? historicalArchivesKnownSources,
     AttachmentArchiveLocationState? attachmentArchiveLocation,
-    AttachmentArchiveRelocationProgress? attachmentArchiveRelocation,
-    bool attachmentArchiveRelocationEnabled = false,
   }) async {
     return spec.when(
       historicalArchivesOverview: () => ref
@@ -60,8 +58,6 @@ class SettingsCassetteCoordinator extends _$SettingsCassetteCoordinator {
           .resolve(
             cassetteIndex: cassetteIndex,
             location: attachmentArchiveLocation!,
-            relocation: attachmentArchiveRelocation,
-            relocationEnabled: attachmentArchiveRelocationEnabled,
           ),
     );
   }